refactor: rename ui components to follow kebab-case
