+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Image from 'next/image';
 
-import Button from '@/components/ui/Button';
+import Button from '@/components/ui/button';
 import Link from 'next/link';
 
 import { TBSVideoPic, joinUsBGPic } from '@/assets/index';
