+++ src/components/footer/newsletter.tsx
@@ -5,8 +5,8 @@ import { FormValues } from '@/types';
 
 import { emailIcon } from '@/assets';
 
-import Button from '@/components/ui/Button';
-import Input from '@/components/ui/Input';
+import Button from '@/components/ui/button';
+import Input from '@/components/ui/input';
 
 const NewsLetter = () => {
 	const { register } = useForm<FormValues>();
